requirements.txt
blinker==1.9.0
click==8.1.8
Deprecated==1.2.18
diskcache==5.6.3
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.12
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
llama_cpp_python==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.5
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.1
rich==13.9.4
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
